rename CreateUserTest, LoginUser, EditUserData, CreateOrder
