+++ src/test/java/EditUserDataTest.java
@@ -6,8 +6,10 @@ import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
 
 import static io.restassured.RestAssured.given;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.notNullValue;
 
-public class CreateOrder {
+public class EditUserDataTest {
     private String login;
     private String accessToken;
 
@@ -46,66 +48,51 @@ public class CreateOrder {
     }
 
     @Test
-    @DisplayName("Auth user create order")
-    public void userCreateOrder() {
-        String json = String.format("{\"ingredients\": [\"61c0c5a71d1f82001bdaaa6d\",\"61c0c5a71d1f82001bdaaa6f\"]}");
+    @DisplayName("Edit user name data")
+    public void editUserNameData(){
+        String json = String.format("{\"email\":\"%s\",\"name\":\"Username2\"}", login);
         given()
                 .header("Content-type", "application/json")
                 .header("Authorization", accessToken)
                 .and()
                 .body(json)
                 .when()
-                .post("/api/orders")
+                .patch("/api/auth/user ")
                 .then().assertThat()
                 .and()
                 .statusCode(200);
     }
 
     @Test
-    @DisplayName("Unauthorized user create order")
-    public void unauthUserCreateOrder() {
-        String json = String.format("{\"ingredients\": [\"61c0c5a71d1f82001bdaaa6d\",\"61c0c5a71d1f82001bdaaa6f\"]}");
+    @DisplayName("Edit user email data")
+    public void editUserEmailData(){
+        login = String.format("a%s@gmail.com", java.util.UUID.randomUUID());
+        String json = String.format("{\"email\":\"%s\",\"name\":\"Username\"}", login);
         given()
                 .header("Content-type", "application/json")
+                .header("Authorization", accessToken)
                 .and()
                 .body(json)
                 .when()
-                .post("/api/orders")
+                .patch("/api/auth/user ")
                 .then().assertThat()
                 .and()
                 .statusCode(200);
     }
 
     @Test
-    @DisplayName("User create order without ingredients")
-    public void userCreateOrderWithoutIngredients() {
-        String json = String.format("{\"ingredients\": []}");
+    @DisplayName("Edit unauthorized user data")
+    public void editUserDataNotAuth(){
+        String json = String.format("{\"email\":\"%s\",\"name\":\"Username2\"}", login);
         given()
                 .header("Content-type", "application/json")
-                .header("Authorization", accessToken)
                 .and()
                 .body(json)
                 .when()
-                .post("/api/orders")
-                .then().assertThat()
+                .patch("/api/auth/user ")
+                .then().assertThat().body("message", equalTo("You should be authorised"))
                 .and()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("User create order with invalid ingredients hash")
-    public void userCreateOrderWithInvalidHash() {
-        String json = String.format("{\"ingredients\": [\"61c0c5a71d1f82001bdaa+++\",\"61c0c5a71d1f82001bdaa___\"]}");
-        given()
-                .header("Content-type", "application/json")
-                .header("Authorization", accessToken)
-                .and()
-                .body(json)
-                .when()
-                .post("/api/orders")
-                .then().assertThat()
-                .and()
-                .statusCode(500);
+                .statusCode(401);
     }
 
     @After
