+++ src/test/java/LoginUserTest.java
@@ -5,16 +5,13 @@ import io.restassured.RestAssured;
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 
-public class LoginUser {
+public class LoginUserTest {
     private String login;
 
     @Before
